Restructured protobuf
