+++ src/__tests__/ProtobufDecoder.test.ts
@@ -1,7 +1,7 @@
 // Copyright 2019 Cognite AS
 
 const fs = require('fs');
-import ProtobufDecoder from '../ProtobufDecoder';
+import ProtobufDecoder from '../parsers/protobuf/ProtobufDecoder';
 import * as TestScene from './fixtures/test_scene.json';
 
 describe('Protobufdecoder', () => {
