+++ src/parsers/protobuf/ProtobufDecoder.ts
